tons of fixes, added base style

+++ transform-document.sh
@@ -8,13 +8,16 @@ if [ ! -f cl-ppcre.sbcl ]; then
   sbcl --noinform --no-userinit --eval "(require :asdf)" --load build-ppcre-core.lisp
 fi
 
+if [ -z $1 ]; then
+  echo "Usage: $0 SOURCEFILE"
+  exit 1
+fi
+
 rm -f tmp
 
 cat "$1" | ./cl-ppcre.sbcl --no-userinit --load syntax.lisp --noprint --noinform > tmp
 
-mv tmp "$1"
-
 ### stx2any
 
-stx2any -thtml "$1" > "$1.html"
+stx2any -thtml --table-of-contents on --numbering on tmp > "$1.html"
